cleanup: compilation
- cleanup: compilation warnings (#import)
- cleanup: compilation warnings (noexcept)
- cleanup: compilation warnings (std::unary_function)
- cleanup: compilation warnings (std::binary_function)
- cleanup: compilation warnings (std::bind1st)
- cleanup: compilation warnings (std::bind2nd)
- cleanup: compilation warnings (std::mem_fun)
- cleanup: compilation warnings (std::mem_fun_ref)
- cleanup: compilation warnings (std::ptr_fun)
- cleanup: compilation warnings (std::auto_ptr)
- cleanup: compilation warnings (mem functions vs non trivial copy)
- cleanup: compilation warnings (misc)
- cleanup: compilation warnings (std::iterator)
- cleanup: compilation warnings (std::random_shuffle)
- fix: compilation error (clang template instanciation)

+++ src/torrent/utils/directory_events.cc
@@ -140,7 +140,7 @@ directory_events::event_read() {
   char buffer[2048];
   int result = ::read(m_fileDesc, buffer, 2048);
 
-  if (result < sizeof(struct inotify_event))
+  if (result < static_cast<ssize_t>(sizeof(struct inotify_event)))
     return;
 
   struct inotify_event* event = (struct inotify_event*)buffer;
